+++ zerver/lib/home.py
@@ -8,7 +8,6 @@ from django.http import HttpRequest
 from django.utils import translation
 from two_factor.utils import default_device
 
-from version import ZULIP_MERGE_BASE
 from zerver.context_processors import get_apps_page_url
 from zerver.lib.events import do_events_register
 from zerver.lib.i18n import (
@@ -214,8 +213,6 @@ def build_page_params_for_home_page_load(
         # There is no event queue for spectators since
         # events support for spectators is not implemented yet.
         no_event_queue=user_profile is None,
-        # Required for about_zulip.hbs
-        zulip_merge_base=ZULIP_MERGE_BASE,
     )
 
     for field_name in register_ret.keys():